+++ zia/services/urlfilteringpolicies/urlfilteringpolicies.go
@@ -29,6 +29,8 @@ type URLFilteringRule struct {
 	// List of URL categories for which rule must be applied
 	URLCategories []string `json:"urlCategories"`
 
+	UserRiskScoreLevels []string `json:"userRiskScoreLevels,omitempty"`
+
 	// Rule State
 	State string `json:"state,omitempty"`
 
@@ -176,11 +178,24 @@ func (service *Service) Create(ruleID *URLFilteringRule) (*URLFilteringRule, err
 	return createdURLFilteringRule, nil
 }
 
-func (service *Service) Update(ruleID int, rules *URLFilteringRule) (*URLFilteringRule, *http.Response, error) {
-	// Add debug log to print the rules object
-	service.Client.Logger.Printf("[DEBUG] Updating URL Filtering Rule with ID %d: %+v", ruleID, rules)
-
-	resp, err := service.Client.UpdateWithPut(fmt.Sprintf("%s/%d", urlFilteringPoliciesEndpoint, ruleID), *rules)
+func (service *Service) Update(ruleID int, rule *URLFilteringRule) (*URLFilteringRule, *http.Response, error) {
+	// Add debug log to print the rule object
+	service.Client.Logger.Printf("[DEBUG] Updating URL Filtering Rule with ID %d: %+v", ruleID, rule)
+	if rule.CBIProfile.ID == "" || rule.CBIProfileID == 0 {
+		// If CBIProfile object is empty, fetch it using GetByName as Get by ID is not currently returnign the full CBIProfile object with the uuid ID
+		var urlFilteringPolicies []URLFilteringRule
+		err := common.ReadAllPages(service.Client, urlFilteringPoliciesEndpoint, &urlFilteringPolicies)
+		if err != nil {
+			return nil, nil, err
+		}
+		for _, urlFilteringPolicy := range urlFilteringPolicies {
+			if urlFilteringPolicy.ID == ruleID {
+				rule.CBIProfile = urlFilteringPolicy.CBIProfile
+				rule.CBIProfileID = urlFilteringPolicy.CBIProfileID
+			}
+		}
+	}
+	resp, err := service.Client.UpdateWithPut(fmt.Sprintf("%s/%d", urlFilteringPoliciesEndpoint, ruleID), *rule)
 	if err != nil {
 		return nil, nil, err
 	}